patches: Re-add `UnwindRegistersSave.S` RIP patch
The previous patch (removed in 7b56f9b17de72d0e237e4af2bf72635c48389f3c)
was erroneous in its implementation but its intent was valid: when
saving registers in unikraft libunwind should take the return address of
the previous stack frame as the IP of the throwing function.
This patch implements this functionality correctly.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>